Creating of func for create route Make migration

+++ Server/Taxi/route/models.py
@@ -9,5 +9,5 @@ class Route(models.Model):
     coordinate_end = models.CharField('Конечные координаты', max_length = 100)
     address_start = models.CharField('Начальный адрес', max_length = 100)
     address_end = models.CharField('Конечный адрес', max_length = 100)
-    distance = models.IntegerField('Дистанция, км')
+    distance = models.FloatField('Дистанция, км')
     map = models.TextField('Карта, html')
